+++ api.js
@@ -1,22 +1,4 @@
-const rp = require('request-promise');
+const api = require('./helpers/apiHelper');
+const { KEEP_ALIVE, BACKEND_URL } = require('./config/global');
 
-const baseUrl = process.env.BACKEND_URL || 'http://localhost:3030/';
-const { KEEP_ALIVE } = process.env;
-
-const api = (req, { json = true } = {}) => {
-	const headers = {};
-	if (req && req.cookies && req.cookies.jwt) {
-		headers.Authorization = (req.cookies.jwt.startsWith('Bearer ') ? '' : 'Bearer ') + req.cookies.jwt;
-	}
-	if (KEEP_ALIVE) {
-		headers.Connection = 'Keep-Alive';
-	}
-
-	return rp.defaults({
-		baseUrl,
-		json,
-		headers,
-	});
-};
-
-module.exports = api;
+module.exports = api(BACKEND_URL, { keepAlive: KEEP_ALIVE });
