+++ helpers/apiHelper.js
@@ -1,0 +1,19 @@
+const rp = require('request-promise');
+
+const api = (baseUrl, { keepAlive = false } = {}) => (req, { json = true } = {}) => {
+	const headers = {};
+	if (req && req.cookies && req.cookies.jwt) {
+		headers.Authorization = (req.cookies.jwt.startsWith('Bearer ') ? '' : 'Bearer ') + req.cookies.jwt;
+	}
+	if (keepAlive) {
+		headers.Connection = 'Keep-Alive';
+	}
+
+	return rp.defaults({
+		baseUrl,
+		json,
+		headers,
+	});
+};
+
+module.exports = api;
